rename initialize -> diContainer
